requirements.txt
slack-bolt==1.18.1
python-dotenv==1.0.0
requests==2.31.0 